3DS.py initial implementation (v0.0.1)
Initial commit for the 3DS xml generator. It's a lightly modified version of the NDS one, to the point where I should really just make a shared function library for a lot of this stuff. (Which is why this isn't even 0.1 yet.)

It does work in my testing, though I'm sure there's bugs. To go along with the change there's some new constants and the readme has been updated as well.